requirements.txt
numpy==1.26.3
torch==1.11.0+cu115
torchvision==0.12.0+cu115
Pillow==10.1.0
matplotlib==3.8.1
seaborn==0.13.2
torchviz==0.0.2
